Fix issues with unused namespace detection
Adresses issue #1. Fixes a bug which caused unused namespaces to not be
removed, if no other namespace was used. Fixes a bug which caused
namespaces to be removed, which have been in used by attributes.

Adds a comment to the README.md that namespace removal will *not*
consider namespaces which are only used for a certain sub-tree of the
document.